version modular del semaforo
