+++ Tps/Tp1/Ejercicio1/src/main.c
@@ -29,117 +29,10 @@ You should have received a copy of the GNU General Public License
 ------------------------------------------------------------------------------*/
 
 #include "main.h"
-#include "sAPI.h"
+#include "Semaphore.h"
 
 
 // TODO(elsuizo:tratar de hacer que el led verde titile 2 seg antes de pasar a amarillo)-->Thu 05 May 2016 05:20:02 PM ART
-
-/*------------------------------------------------------------------------------
-                           State diagram 
-------------------------------------------------------------------------------
-+-------+             +-------+
-| RED   |             | GREEN |
-|       +----+  +-----+       |
-+----+--+    |  |     +---+---+
-     ^       |  |         ^
-     |     +-+--+--+      |
-     |     |       |      |
-     +-----+YELLOW +------+
-           |       |
-           +-------+
-------------------------------------------------------------------------------*/
-bool_t timeout;
-
-delay_t delayRED_GREEN;
-delay_t delayBlink;
-delay_t delayYELLOW;
-
-typedef enum{GREEN, YELLOW, RED}state; // define the states
-//  
-state next_state;
-
-bool_t from_RED = FALSE;
-bool_t from_GREEN = FALSE;
-/*------------------------------------------------------------------------------
-                            init_FSM()
-------------------------------------------------------------------------------*/
-void init_FSM()
-{
-   next_state = RED;
-} 
-
-state do_state_YELLOW()
-{
-   while(!delayRead(&delayYELLOW))
-   {
-		digitalWrite(LED3, OFF);
-		digitalWrite(LED1, OFF);
-		digitalWrite(LED2, ON);
-
-   }
-   if(from_GREEN)
-   {
-      from_GREEN = FALSE;
-      timeout = FALSE;
-      return RED;
-   }
-   if(from_RED)
-   {
-      from_RED = FALSE;
-      timeout = FALSE;
-      return GREEN;
-   }
-}
-
-state do_state_GREEN()
-{
-   while(!delayRead(&delayRED_GREEN))
-   {
-		digitalWrite(LED3, ON);
-	   digitalWrite(LED1, OFF);
-		digitalWrite(LED2, OFF);
-   }
-   timeout = FALSE;
-   from_GREEN = TRUE;
-   return YELLOW;
-}
-
-state do_state_RED()
-{
-   while(!delayRead(&delayRED_GREEN))
-   {
-		digitalWrite(LED3, OFF);
-		digitalWrite(LED1, ON);
-		digitalWrite(LED2, OFF);
-   }
-   timeout = FALSE;
-   from_RED = TRUE;
-   return YELLOW;
-}
-
-
-/*------------------------------------------------------------------------------
-                           update_FSM()
-/------------------------------------------------------------------------------*/
-void update_FSM()
-{
-   switch(next_state)
-   {
-
-      case YELLOW:
-         next_state = do_state_YELLOW();
-         break;
-
-      case GREEN:
-         next_state = do_state_GREEN();
-         break;
-
-      case RED:
-         next_state = do_state_RED();
-         break;
-
-   }
-}
 
 int main(void)
 {
